+++ mlxtend/feature_selection/tests/test_sequential_feature_selector.py
@@ -9,6 +9,9 @@ import sys
 import numpy as np
 from numpy import nan
 from numpy.testing import assert_almost_equal
+
+import pandas as pd
+
 from sklearn.datasets import load_boston
 from sklearn.datasets import load_iris
 from sklearn.ensemble import RandomForestClassifier
@@ -789,3 +792,87 @@ def test_max_feature_subset_parsimonious():
 
     sfs = sfs.fit(X, y)
     assert sfs.k_feature_idx_ == (5, 10, 11, 12)
+
+
+def test_check_pandas_dataframe_fit():
+
+    iris = load_iris()
+    X = iris.data
+    y = iris.target
+    lr = SoftmaxRegression(random_seed=1)
+    sfs1 = SFS(lr,
+               k_features=2,
+               forward=True,
+               floating=False,
+               scoring='accuracy',
+               cv=0,
+               verbose=0,
+               n_jobs=1)
+
+    df = pd.DataFrame(X, columns=['sepal len', 'sepal width',
+                                  'petal len', 'petal width'])
+
+    sfs1 = sfs1.fit(X, y)
+    assert sfs1.k_feature_idx_ == (1, 3)
+    assert sfs1.k_feature_names_ == ('1', '3')
+    assert sfs1.subsets_[2]['feature_names'] == ('1', '3')
+
+    sfs1 = sfs1.fit(df, y)
+    assert sfs1.subsets_[1]['feature_names'] == ('petal width',)
+    assert sfs1.subsets_[2]['feature_names'] == ('sepal width', 'petal width')
+    assert sfs1.subsets_[1]['feature_idx'] == (3,)
+    assert sfs1.subsets_[2]['feature_idx'] == (1, 3)
+    assert sfs1.k_feature_idx_ == (1, 3)
+    assert sfs1.k_feature_names_ == ('sepal width', 'petal width')
+
+    sfs1._TESTING_INTERRUPT_MODE = True
+    out = sfs1.fit(df, y)
+    assert len(out.subsets_.keys()) > 0
+    assert sfs1.interrupted_
+    assert sfs1.subsets_[1]['feature_names'] == ('petal width',)
+    assert sfs1.k_feature_idx_ == (3,)
+    assert sfs1.k_feature_names_ == ('petal width',)
+
+
+def test_check_pandas_dataframe_transform():
+    iris = load_iris()
+    X = iris.data
+    y = iris.target
+    lr = SoftmaxRegression(random_seed=1)
+    sfs1 = SFS(lr,
+               k_features=2,
+               forward=True,
+               floating=False,
+               scoring='accuracy',
+               cv=0,
+               verbose=0,
+               n_jobs=1)
+
+    df = pd.DataFrame(X, columns=['sepal length', 'sepal width',
+                                  'petal length', 'petal width'])
+    sfs1 = sfs1.fit(df, y)
+    assert sfs1.k_feature_idx_ == (1, 3)
+    assert (150, 2) == sfs1.transform(df).shape
+
+
+def test_custom_feature_names():
+
+    iris = load_iris()
+    X = iris.data
+    y = iris.target
+    lr = SoftmaxRegression(random_seed=1)
+    sfs1 = SFS(lr,
+               k_features=2,
+               forward=True,
+               floating=False,
+               scoring='accuracy',
+               cv=0,
+               verbose=0,
+               n_jobs=1)
+
+    sfs1 = sfs1.fit(X, y, custom_feature_names=(
+          'sepal length', 'sepal width', 'petal length', 'petal width'))
+    assert sfs1.k_feature_idx_ == (1, 3)
+    assert sfs1.k_feature_names_ == ('sepal width', 'petal width')
+    assert sfs1.subsets_[2]['feature_names'] == ('sepal width',
+                                                 'petal width')